pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dracula-manager"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12",
    "rich>=13",
    "PyGithub>=2.3",
    "gitpython>=3.1",
    "pydantic>=2",
    "json5>=0.9",
    "tenacity>=8",
]

[dependency-groups]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "bandit>=1.7",
    "pre-commit>=3",
    "pytest>=8",
    "pytest-mock>=3",
    "pytest-cov>=5",
    "freezegun>=1",
]

[project.scripts]
dracula = "dracula_manager.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.bandit]
skips = ["B404", "B603", "B607", "B110"]

[tool.uv]
package = true
